Fix #299, creating utility file for tests.
